Update the transactions/JMS/SE tree
* Move to Jakarta
* Move CF creation to common part
* Allow config to be overridden by environment variables
* Make success/failure return codes more consistent to permit testing
* Update READMEs and other text files
* Removed some of the verbosity in logging
* Make version-agnostic
