+++ transactions/JMS/SE/com/ibm/mq/samples/jms/SimpleJmsTransaction.java
@@ -1,5 +1,5 @@
 /*
- * (c) Copyright IBM Corporation 2020, 2023
+ * (c) Copyright IBM Corporation 2020, 2026
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,26 +16,26 @@
 
 package com.ibm.mq.samples.jms;
 
-import com.ibm.msg.client.jms.JmsConnectionFactory;
-import com.ibm.msg.client.jms.JmsFactoryFactory;
-import com.ibm.msg.client.wmq.WMQConstants;
 import java.util.Random;
-import javax.jms.Destination;
-import javax.jms.JMSContext;
-import javax.jms.JMSException;
-import javax.jms.JMSProducer;
-import javax.jms.TextMessage;
+
+import com.ibm.msg.client.jakarta.jms.JmsConnectionFactory;
+
+import jakarta.jms.Destination;
+import jakarta.jms.JMSContext;
+import jakarta.jms.JMSException;
+import jakarta.jms.JMSProducer;
+import jakarta.jms.JMSRuntimeException;
+import jakarta.jms.TextMessage;
 
 /**
- * A minimal and simple application for Point-to-point messaging with a transaction.
+ * A minimal and simple application for Point-to-point messaging with a
+ * transaction.
  *
- * Application makes use of fixed literals, any customisations will require
- * re-compilation of this source file. Application assumes that the named queue
- * is empty prior to a run.
+ * Application assumes that the named queue is empty prior to a run.
  *
  * Notes:
  *
- * API type: JMS API (v2.0, simplified domain)
+ * API type: Jakarta JMS API
  *
  * Messaging domain: Point-to-point
  *
@@ -46,19 +46,10 @@ import javax.jms.TextMessage;
  * JNDI in use: No
  *
  */
-public class simpleJmsTransaction {
+public class SimpleJmsTransaction {
 
   // System exit status value (assume unset value to be 1)
   private static int status = 1;
-
-  // Create variables for the connection to MQ
-  private static final String HOST = "localhost"; // Host name or IP address
-  private static final int PORT = 1414; // Listener port for your queue manager
-  private static final String CHANNEL = "DEV.APP.SVRCONN"; // Channel name
-  private static final String QMGR = "QM1"; // Queue manager name
-  private static final String APP_USER = "app"; // User name that application uses to connect to MQ
-  private static final String APP_PASSWORD = "passw0rd"; // Password that the application uses to connect to MQ
-  private static final String QUEUE_NAME = "DEV.QUEUE.1"; // Queue that the application uses to put and get messages to and from
 
   /**
    * Main method
@@ -71,11 +62,13 @@ public class simpleJmsTransaction {
     Destination destination = null;
     JMSProducer producer = null;
 
-    Random rand = new Random(); //instance of random class
+    Random rand = new Random(); // instance of random class
     int upperbound = 999;
 
     // Defining a specific exception for when rollback is occuring
     class PutTransactionRollbackException extends Exception {
+
+      private static final long serialVersionUID = 1L;
 
       public PutTransactionRollbackException(String s) {
         // Call constructor of parent Exception
@@ -84,74 +77,58 @@ public class simpleJmsTransaction {
     }
 
     try {
-      // Create a connection factory
-      JmsFactoryFactory ff = JmsFactoryFactory.getInstance(
-        WMQConstants.WMQ_PROVIDER
-      );
-      JmsConnectionFactory cf = ff.createConnectionFactory();
-
-      // Set the properties
-      cf.setStringProperty(WMQConstants.WMQ_HOST_NAME, HOST);
-      cf.setIntProperty(WMQConstants.WMQ_PORT, PORT);
-      cf.setStringProperty(WMQConstants.WMQ_CHANNEL, CHANNEL);
-      cf.setIntProperty(
-        WMQConstants.WMQ_CONNECTION_MODE,
-        WMQConstants.WMQ_CM_CLIENT
-      );
-      cf.setStringProperty(WMQConstants.WMQ_QUEUE_MANAGER, QMGR);
-      cf.setStringProperty(
-        WMQConstants.WMQ_APPLICATIONNAME,
-        "simpleJmsTransaction (JMS)"
-      );
-      cf.setBooleanProperty(WMQConstants.USER_AUTHENTICATION_MQCSP, true);
-      cf.setStringProperty(WMQConstants.USERID, APP_USER);
-      cf.setStringProperty(WMQConstants.PASSWORD, APP_PASSWORD);
+      JmsConnectionFactory cf = Common.createCF();
 
       // Create JMS objects
+      context = cf.createContext(JMSContext.SESSION_TRANSACTED);
+      destination = context.createQueue("queue:///" + Common.QUEUE_NAME);
 
       int uniqueNumber = rand.nextInt(upperbound);
       System.out.println("Your number is: " + uniqueNumber);
-
-      context = cf.createContext(JMSContext.SESSION_TRANSACTED);
-      destination = context.createQueue("queue:///" + QUEUE_NAME);
-
-      TextMessage message = context.createTextMessage(
-        "Your lucky number today is " + uniqueNumber
-      );
+      TextMessage message = context.createTextMessage("Your lucky number today is " + uniqueNumber);
 
       producer = context.createProducer();
-
       producer.send(destination, message);
       System.out.println("Sent message:\n" + message);
 
       try {
-        // Sleep for 15 seconds before publishing the next event
-        Thread.sleep(15000);
+        // Sleep for 5 seconds before publishing the next event
+        Thread.sleep(5000);
       } catch (InterruptedException e) {
-        System.out.println("wait interrupt 15 sec");
+        System.out.println("wait interrupt 5 sec");
       }
       if (uniqueNumber % 2 == 0) {
-        throw new PutTransactionRollbackException(
-          "Lucky number was even, rolling back"
-        );
+        throw new PutTransactionRollbackException("Lucky number was even, rolling back");
       }
       context.commit();
 
       recordSuccess();
-    } catch (JMSException jmsex) {
-      context.rollback();
+    } catch (JMSException | JMSRuntimeException jmsex) {
+      if (context != null) {
+        context.rollback();
+      }
       recordFailure(jmsex);
       System.out.println("JMSEX ");
       jmsex.printStackTrace();
     } catch (PutTransactionRollbackException ptsex) {
       context.rollback();
       recordFailure(ptsex);
-      System.out.println(
-        "Rollback was successful, message was not put to queue"
-      );
+      System.out.println("Rollback was successful, message was not put to queue");
+      // This is an expected condition to reset status to OK
+      status = 0;
     } catch (Exception ex) {
       System.out.println("EX ");
       ex.printStackTrace();
+    }
+
+    if (context != null) {
+      try {
+        context.stop();
+        context.close();
+
+      } catch (Exception e) {
+        // ignore the exception
+      }
     }
     System.exit(status);
   } // end main()
@@ -171,6 +148,8 @@ public class simpleJmsTransaction {
    * @param ex
    */
   private static void recordFailure(Exception ex) {
+    status = 0;
+
     if (ex != null) {
       if (ex instanceof JMSException) {
         processJMSException((JMSException) ex);
@@ -179,7 +158,6 @@ public class simpleJmsTransaction {
       }
     }
     System.out.println("FAILURE");
-    status = -1;
     return;
   }
